move: index.tsx to YearMonthPicker.tsx
